Fix: 이메일 체크 로직 수정
