+++ src/main/java/what/what2eat/domain/auth/controller/CommonAuthController.java
@@ -6,6 +6,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import what.what2eat.domain.auth.entity.Provider;
 import what.what2eat.domain.auth.service.CommonAuthService;
 import what.what2eat.global.response.ApiResponse;
 import what.what2eat.global.response.ResponseCode;
@@ -26,15 +27,8 @@ public class CommonAuthController {
         return ResponseEntity.ok(ApiResponse.of(ResponseCode.SUCCESS));
     }
 
-    @PostMapping("/check-email")
-    @Operation(summary = "이메일 중복 체크", description = "이메일 중복 체크를 처리합니다. 이메일을 제공해야 합니다. \n 응답코드에 따른 결과값은 포스트맨 API 명세서를 참고 부탁드립니다.")
-    public ResponseEntity<ApiResponse<Boolean>> checkEmail(@RequestParam String userEmail) {
-        commonAuthService.validateMember(userEmail);
-
-        return ResponseEntity.ok(ApiResponse.of(ResponseCode.SUCCESS));
-    }
-
     @DeleteMapping("")
+    @Operation(summary = "공통 회원탈퇴", description = "회원탈퇴를 처리합니다. Authorization 헤더에 accessToken을 첨부해서 요청하시면 됩니다. \n 응답 코드에 따른 자세한 결과는 PostMan API 명세서를 참고 부탁드립니다.")
     public ResponseEntity<ApiResponse<ResponseCode>> withdrawal(HttpServletRequest request) {
         commonAuthService.delete(request);
