+++ src/main/java/what/what2eat/domain/auth/service/LocalAuthService.java
@@ -31,7 +31,6 @@ public class LocalAuthService {
     private final AuthRepository authRepository;
     private final PasswordEncoder passwordEncoder;
     private final JwtProvider jwtProvider;
-    private final CommonAuthService commonAuthService;
 
     // 회원가입 =>
     public void signUp(LocalAuthRequestDTO.SignUpRequestDTO request) {
@@ -54,7 +53,7 @@ public class LocalAuthService {
     public LocalAuthResponseDTO.LoginResponseDTO login(LocalAuthRequestDTO.LoginRequestDTO request) throws Exception {
 
         // 유효성 검사
-        commonAuthService.validateMember(request.getUserEmail());
+        validateMember(request.getUserEmail());
 
         try {
             // 인증 시도
@@ -92,5 +91,12 @@ public class LocalAuthService {
     }
 
 
+    // 로그인시
+    public void validateMember(String userEmail) {
+        Boolean isExist = authRepository.existsByUserEmailAndUserStatusAndProvider(userEmail, UserStatus.ACTIVE, Provider.LOCAL);
 
+        // 이메일이 존재하지 않으면 404 에러 반환
+        if(!isExist) throw new MemberException(AuthErrorCode.USER_NOT_FOUND);
+
+    }
 }
